Deprecate range type (was unused)
It's syntax was questionable because the colon (':') token is
overused already and causes conflicts and confusion.
Also it wasn't used anywhere and removing this makes the language simpler.
The range type idea may be revised in the future as an extension,
it was inspired by some other languages that does have them to quickly
access array and multidimensional sub ranges.