Introduce the resource pool class (used internally by 'io' module)
This is an utility class that provides a mechanism for allocating
resources from a pool associated with an unique 'id'.

It is used internally by the 'io' module to properly handle dangling files.
The resource pool returns a new unique id for every new allocated object.
This id is composed by an index in the lower bits and a generation counter
in the higher bits. The generation counter is incremented every new allocation
for the same index.

It's main use is to give the possibility to detect dangling accesses,
that is trying to access an object which no longer exists or that
has been reused. These situations is not possible to detect when using
just raw pointers.